removed debug simulation code from map file into its own file to seperate concerns.
