+++ frontend/src/app/components/map/map/map.html
@@ -1,4 +1,4 @@
-<!-- <div class="map-container">
+<div class="map-container">
   <div *ngIf="isLoading" class="loading-overlay">
     <div class="loading-spinner">Loading map...</div>
   </div>
@@ -8,6 +8,4 @@
   </div>
   
   <div #mapContainer id="mapContainer" style="width: 100%; height: 400px;"></div>
-</div> -->
-
-<div #mapContainer style="width: 100%; height: 400px;"></div>
+</div>
